# [Tests] Speed up tests by using paratest