--- a/composer.json
+++ b/composer.json
@@ -34,6 +34,7 @@
         "twig/twig": "^1.35.4 | ~2.5"
     },
     "require-dev": {
+        "brianium/paratest": "^2.2",
         "friendsofphp/php-cs-fixer": "~2.7.5",
         "phpunit/phpunit": "^7.0",
         "matthiasnoback/symfony-dependency-injection-test": "~3.0",
@@ -66,10 +67,12 @@
         }
     },
     "scripts": {
-        "fix-cs": "@php ./vendor/bin/php-cs-fixer fix -v --show-progress=estimating",
+        "fix-cs": "php-cs-fixer fix -v --show-progress=estimating",
+        "unit": "paratest -p half -c phpunit.xml --runner=WrapperRunner",
+        "integration": "paratest -p half -c phpunit-integration-legacy.xml --runner=WrapperRunner",
         "test": [
-            "phpunit -c phpunit.xml",
-            "phpunit -c phpunit-integration-legacy.xml"
+            "@unit",
+            "@integration"
         ]
     },
     "scripts-descriptions": {