--- a/phpunit-integration-legacy-elasticsearch.xml
+++ b/phpunit-integration-legacy-elasticsearch.xml
@@ -54,7 +54,6 @@
           <file>eZ/Publish/API/Repository/Tests/SearchServiceLocationTest.php</file>
           <file>eZ/Publish/API/Repository/Tests/SearchServiceAuthorizationTest.php</file>
           <file>eZ/Publish/API/Repository/Tests/SearchEngineIndexingTest.php</file>
-          <file>eZ/Publish/API/Repository/Tests/LimitationTest.php</file>
           <file>eZ/Publish/API/Repository/Tests/FieldTypeServiceTest.php</file>
         </testsuite>
     </testsuites>