| Question           | Answer
| ------------------ | ------------------
| **JIRA issue**     | N/A
| **Bug/Improvement**| yes/no
| **New feature**    | yes/no
| **Target version** | `7.x` _(paratest requires php7)_
| **BC breaks**      | no
| **Tests pass**     | yes/no
| **Doc needed**     | no

I don't expect this to pass with mysql/postgres/sorl, as we should setup different databases for the different processes using [test token](https://github.com/paratestphp/paratest#test-token).


**TODO**:
- [x] Implement feature / fix a bug.
- [x] Implement tests.
- [x] Fix new code according to Coding Standards (`$ composer fix-cs`).
- [x] Ask for Code Review.