@andrerom 2018-10-19
As expected it fails on postgres/mysql/solr/elastic, BUT it fails _much_ faster... 🤣