@andrerom 2018-10-19
As Travis only has 2 cores, it does not seem to be worth it there. While runs with mysql and postgres  passes now, they are slower probably due to 2 php processes fighting db process for cpu. unit tests however are about 10-20sec faster.

Locally I have the following results:
- `composer tests`: 36sec => 8sec
- `composer integration-tests`: 1m 25sec => 26sec

Any opinion on if we should have this? And for travis, should we use it for unit tests? _(mainly to make sure phpunit config is valid as paratest is strict on this, it will somewhat complicating .travis.yml tough if we want to skip it for integration tests)_