@andrerom 2018-11-27
Ready for review now, but see inline discussion in regards to changes in 896c49b which should rather be somehow applied to 6.7 and up.